Create backend configuration to share Terraform state
Using Terraform between devices or with teams will require the use of
same terraform.tfstate file. That file could be committed into a
private CSV but a better option is to use a Terraform backend.
It will keep the state into the selected provider preventing also
leakage of sensitive data.

https://www.terraform.io/docs/backends/index.html